Rename test/extended/util/cloud to test/extended/util/cluster
It already contains non-cloud-specific network info, and will soon
contain more.
