+++ cmd/openshift-tests/provider.go
@@ -13,7 +13,7 @@ import (
 	e2e "k8s.io/kubernetes/test/e2e/framework"
 
 	exutil "github.com/openshift/origin/test/extended/util"
-	exutilcloud "github.com/openshift/origin/test/extended/util/cloud"
+	exutilcluster "github.com/openshift/origin/test/extended/util/cluster"
 
 	// Initialize baremetal as a provider
 	_ "github.com/openshift/origin/test/extended/util/baremetal"
@@ -31,7 +31,7 @@ import (
 
 type TestNameMatchesFunc func(name string) bool
 
-func initializeTestFramework(context *e2e.TestContextType, config *exutilcloud.ClusterConfiguration, dryRun bool) error {
+func initializeTestFramework(context *e2e.TestContextType, config *exutilcluster.ClusterConfiguration, dryRun bool) error {
 	// update context with loaded config
 	context.Provider = config.ProviderName
 	context.CloudConfig = e2e.CloudConfig{
@@ -67,7 +67,7 @@ func initializeTestFramework(context *e2e.TestContextType, config *exutilcloud.C
 	return nil
 }
 
-func getProviderMatchFn(config *exutilcloud.ClusterConfiguration) TestNameMatchesFunc {
+func getProviderMatchFn(config *exutilcluster.ClusterConfiguration) TestNameMatchesFunc {
 	// given the configuration we have loaded, skip tests that our provider, disconnected status,
 	// or our network plugin should exclude
 	var skips []string
@@ -90,19 +90,19 @@ func getProviderMatchFn(config *exutilcloud.ClusterConfiguration) TestNameMatche
 	return matchFn
 }
 
-func decodeProvider(provider string, dryRun, discover bool, clusterState *exutilcloud.ClusterState) (*exutilcloud.ClusterConfiguration, error) {
+func decodeProvider(provider string, dryRun, discover bool, clusterState *exutilcluster.ClusterState) (*exutilcluster.ClusterConfiguration, error) {
 	switch provider {
 	case "none":
-		return &exutilcloud.ClusterConfiguration{ProviderName: "skeleton"}, nil
+		return &exutilcluster.ClusterConfiguration{ProviderName: "skeleton"}, nil
 
 	case "":
 		if _, ok := os.LookupEnv("KUBE_SSH_USER"); ok {
 			if _, ok := os.LookupEnv("LOCAL_SSH_KEY"); ok {
-				return &exutilcloud.ClusterConfiguration{ProviderName: "local"}, nil
+				return &exutilcluster.ClusterConfiguration{ProviderName: "local"}, nil
 			}
 		}
 		if dryRun {
-			return &exutilcloud.ClusterConfiguration{ProviderName: "skeleton"}, nil
+			return &exutilcluster.ClusterConfiguration{ProviderName: "skeleton"}, nil
 		}
 		fallthrough
 
@@ -112,12 +112,12 @@ func decodeProvider(provider string, dryRun, discover bool, clusterState *exutil
 			if err != nil {
 				return nil, err
 			}
-			clusterState, err = exutilcloud.DiscoverClusterState(clientConfig)
+			clusterState, err = exutilcluster.DiscoverClusterState(clientConfig)
 			if err != nil {
 				return nil, err
 			}
 		}
-		config, err := exutilcloud.LoadConfig(clusterState)
+		config, err := exutilcluster.LoadConfig(clusterState)
 		if err != nil {
 			return nil, err
 		}
@@ -140,19 +140,19 @@ func decodeProvider(provider string, dryRun, discover bool, clusterState *exutil
 		}
 		// attempt to load the default config, then overwrite with any values from the passed
 		// object that can be overridden
-		var config *exutilcloud.ClusterConfiguration
+		var config *exutilcluster.ClusterConfiguration
 		if discover {
 			if clusterState == nil {
 				if clientConfig, err := e2e.LoadConfig(true); err == nil {
-					clusterState, _ = exutilcloud.DiscoverClusterState(clientConfig)
+					clusterState, _ = exutilcluster.DiscoverClusterState(clientConfig)
 				}
 			}
 			if clusterState != nil {
-				config, _ = exutilcloud.LoadConfig(clusterState)
+				config, _ = exutilcluster.LoadConfig(clusterState)
 			}
 		}
 		if config == nil {
-			config = &exutilcloud.ClusterConfiguration{
+			config = &exutilcluster.ClusterConfiguration{
 				ProviderName: providerInfo.Type,
 				ProjectID:    providerInfo.ProjectID,
 				Region:       providerInfo.Region,
